backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.12
requests==2.31.0
python-dotenv==1.0.0